transfer features from cg2sandbox * add axis aligned bounding box * add lazzy camera * add navigation math * rename trackball * add world-in-hand navigation

+++ include/glowutils/MathMacros.h
@@ -33,10 +33,10 @@ namespace glow
     ((x) < (l) ? (l) : (x) > (u) ? (u) : (x))
 
 #define deg(rad) \
-    ((rad) * 180.0L / PI)
+    static_cast<float>((rad)* 180.0L / PI)
 
 #define rad(deg) \
-    ((deg) * PI / 180.0L)
+    static_cast<float>((deg)* PI / 180.0L)
 
 #define mi(a, b) \
     ((a) < (b) ? (a) : (b))
